+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Added player's killer bounty
+  * Mobs that kill players will not despawn and will drop 4x more experience and loot
 * Fixed diamond generation backport not working
 * Fixed Wandering Trader trades being not stackable